+++ index.js
@@ -1,14 +1,5 @@
 'use strict';
 
-const assert = require('assert'),
-    camelCase = require('camel-case'),
-    fs = require('fs'),
-    mkdirp = require('mkdirp'),
-    path = require('path'),
-    PNGImage = require('png-image'),
-    PNGJSImage = require('pngjs-image'),
-    resembleJS = require('./lib/resemble');
-
 /**
  * image-diff protractor plugin class
  *
@@ -17,8 +8,9 @@ const assert = require('assert'),
  * @param {object} options
  * @param {string} options.baselineFolder Path to the baseline folder
  * @param {string} options.screenshotPath Path to the folder where the screenshots are saved
+ * @param {boolean} options.autoSaveBaseline If no baseline image is found the image is automatically copied to the baselinefolder
  * @param {boolean} options.debug Add some extra logging and always save the image difference (default:false)
- * @param {string} options.formatImageOptions Custom variables for Image Name (default:{tag}-{browserName}-{width}x{height}-dpr-{dpr})
+ * @param {string} options.formatImageName Custom variables for Image Name (default:{tag}-{browserName}-{width}x{height}-dpr-{dpr})
  * @param {boolean} options.disableCSSAnimation Disable all css animations on a page (default:false)
  * @param {boolean} options.nativeWebScreenshot If a native screenshot of a device (complete screenshot) needs to be taken (default:false)
  * @param {boolean} options.blockOutStatusBar  If the statusbar on mobile / tablet needs to blocked out by default
@@ -38,6 +30,14 @@ const assert = require('assert'),
  * @property {int} fullPageScrollTimeout Default timeout to wait after a scroll
  */
 
+const assert = require('assert');
+const camelCase = require('camel-case');
+const fs = require('fs-extra');
+const path = require('path');
+const PNGImage = require('png-image');
+const PNGJSImage = require('pngjs-image');
+const resembleJS = require('./lib/resemble');
+
 class protractorImageComparison {
     constructor(options) {
         assert.ok(options.baselineFolder, 'Image baselineFolder not given.');
@@ -45,9 +45,10 @@ class protractorImageComparison {
 
         this.baselineFolder = path.normalize(options.baselineFolder);
         this.baseFolder = path.normalize(options.screenshotPath);
+        this.autoSaveBaseline = options.autoSaveBaseline || false;
         this.debug = options.debug || false;
-        this.formatString = options.formatImageName || '{tag}-{browserName}-{width}x{height}-dpr-{dpr}';
         this.disableCSSAnimation = options.disableCSSAnimation || false;
+        this.formatString = options.formatImageName || '{tag}-{browserName}-{width}x{height}-dpr-{dpr}';
 
         this.nativeWebScreenshot = options.nativeWebScreenshot ? true : false;
         this.blockOutStatusBar = options.blockOutStatusBar ? true : false;
@@ -56,17 +57,17 @@ class protractorImageComparison {
         this.ignoreColors = options.ignoreColors || false;
 
         // OS offsets
-        let androidOffsets = options.androidOffsets && typeof options.androidOffsets === 'object' ? options.androidOffsets : {},
-            iosOffsets = options.iosOffsets && typeof options.iosOffsets === 'object' ? options.iosOffsets : {},
-            androidDefaultOffsets = {
-                statusBar: 24,
-                addressBar: 56,
-                toolBar: 48
-            },
-            iosDefaultOffsets = {
-                statusBar: 20,
-                addressBar: 44
-            };
+        let androidOffsets = options.androidOffsets && typeof options.androidOffsets === 'object' ? options.androidOffsets : {};
+        let iosOffsets = options.iosOffsets && typeof options.iosOffsets === 'object' ? options.iosOffsets : {};
+        let androidDefaultOffsets = {
+            statusBar: 24,
+            addressBar: 56,
+            toolBar: 48
+        };
+        let iosDefaultOffsets = {
+            statusBar: 20,
+            addressBar: 44
+        };
 
         this.androidOffsets = this._mergeDefaultOptions(androidDefaultOffsets, androidOffsets);
         this.iosOffsets = this._mergeDefaultOptions(iosDefaultOffsets, iosOffsets);
@@ -84,39 +85,36 @@ class protractorImageComparison {
         this.chromeShadowPadding = 6;
 
         this.tempFullScreenFolder = path.join(this.baseFolder, 'tempFullScreen');
-        this.fullPageScrollTimeout = 1000;
+        this.fullPageScrollTimeout = 1500;
 
-        if (!fs.existsSync(this.diffFolder) || !fs.statSync(this.diffFolder).isDirectory()) {
-            mkdirp.sync(this.diffFolder);
-        }
-
-        if (!fs.existsSync(this.baselineFolder) || !fs.statSync(this.baselineFolder).isDirectory()) {
-            mkdirp.sync(this.baselineFolder);
-        }
-
-        if (!fs.existsSync(this.actualFolder) || !fs.statSync(this.actualFolder).isDirectory()) {
-            mkdirp.sync(this.actualFolder);
-        }
-
-        if (!fs.existsSync(this.tempFullScreenFolder) || !fs.statSync(this.tempFullScreenFolder).isDirectory()) {
-            mkdirp.sync(this.tempFullScreenFolder);
-        }
+        fs.ensureDirSync(this.actualFolder);
+        fs.ensureDirSync(this.baselineFolder);
+        fs.ensureDirSync(this.diffFolder);
+        fs.ensureDirSync(this.tempFullScreenFolder);
     }
 
     /**
-     * Checks if image exists as a baseline image
+     * Checks if image exists as a baseline image, if not, create a baseline image if needed
      * @param {string} tag
      * @returns {Promise}
      * @private
      */
     _checkImageExists(tag) {
         return new Promise((resolve, reject) => {
-            fs.access(path.join(this.baselineFolder, this._formatFileName(tag)), fs.F_OK, error => {
+            fs.access(path.join(this.baselineFolder, this._formatFileName(tag)), error => {
                 if (error) {
-                    reject('Image not found, saving current image as new baseline.');
-                } else {
-                    resolve();
+                    if (this.autoSaveBaseline) {
+                        try {
+                            fs.copySync(path.join(this.actualFolder, this._formatFileName(tag)), path.join(this.baselineFolder, this._formatFileName(tag)));
+                            console.log(`\nINFO: Autosaved the image to ${path.join(this.baselineFolder, this._formatFileName(tag))}\n`);
+                        } catch (error) {
+                            reject(`Image could not be copied. The following error was thrown: ${error}`);
+                        }
+                    } else {
+                        reject('Image not found, saving current image as new baseline.');
+                    }
                 }
+                resolve();
             });
         });
     }
@@ -131,29 +129,27 @@ class protractorImageComparison {
      * @private
      */
     _composeAndSaveFullScreenshot(tag, parts, part, image) {
-        const imageHeight = this.fullPageHeight * this.devicePixelRatio;
-        const imageWidth = this.fullPageWidth * this.devicePixelRatio;
+        const imageHeight = (this.fullPageHeight * this.devicePixelRatio) - (this.chromeShadowPadding* this.devicePixelRatio);
+        // const imageWidth = this.fullPageWidth * this.devicePixelRatio; // @todo: changed this for mobile to clientWidth
+        const imageWidth = this.clientWidth * this.devicePixelRatio;
         const height = (this.innerHeight - this.chromeShadowPadding) * (part - 1) * this.devicePixelRatio;
 
-        if (this.debug) {
-            console.log('\n####################################################');
-            console.log('_composeAndSaveFullScreenshot()');
-            console.log('imageHeight = ', imageHeight);
-            console.log('innerHeight = ', this.innerHeight);
-            console.log('imageWidth = ', imageWidth);
-            console.log('height = ', height);
-            console.log('####################################################\n');
-        }
+        // if (this.debug) {
+        //     console.log('\n####################################################');
+        //     console.log('_composeAndSaveFullScreenshot()');
+        //     console.log('imageHeight = ', imageHeight);
+        //     console.log('innerHeight = ', this.innerHeight);
+        //     console.log('imageWidth = ', imageWidth);
+        //     console.log('height = ', height);
+        //     console.log('####################################################\n');
+        // }
 
         let imageOutput = image || null;
 
         if (part === 1) {
             imageOutput = PNGJSImage.createImage(imageWidth, imageHeight);
         } else if (part > parts) {
-            return new Promise(resolve => {
-                imageOutput.writeImageSync(path.join(this.actualFolder, this._formatFileName(tag)));
-                resolve();
-            });
+            return Promise.resolve(imageOutput.writeImageSync(path.join(this.actualFolder, this._formatFileName(tag))));
         }
 
         const stitchImage = PNGJSImage.readImageSync(path.join(this.tempFullScreenFolder, this._formatFileName(`${tag}-${part}`)));
@@ -169,11 +165,11 @@ class protractorImageComparison {
      * @private
      */
     _determineRectangles(element) {
-        let height,
-            rect,
-            width,
-            x,
-            y;
+        let height;
+        let rect;
+        let width;
+        let x;
+        let y;
 
         return element.getSize()
             .then(elementSize => {
@@ -223,8 +219,8 @@ class protractorImageComparison {
      * @private
      */
     _determineImageComparisonPaths(tag) {
-        const imageComparisonPaths = {},
-            tagName = this._formatFileName(tag);
+        const imageComparisonPaths = {};
+        const tagName = this._formatFileName(tag);
 
         imageComparisonPaths['actualImage'] = path.join(this.actualFolder, tagName);
         imageComparisonPaths['baselineImage'] = path.join(this.baselineFolder, tagName);
@@ -236,11 +232,11 @@ class protractorImageComparison {
     /**
      * Compare images against each other
      * @param {string} tag The tag that is used
-     * @param {object} options comparison options
-     * @param {object} options.blockOut blockout with x, y, width and height values
-     * @param {boolean} options.blockOutStatusBar blockout the statusbar yes or no, it will override the global
-     * @param {boolean} options.ignoreAntialiasing compare images an discard anti aliasing
-     * @param {boolean} options.ignoreColors Even though the images are in colour, the comparison wil compare 2 black/white images
+     * @param {object} compareOptions comparison options
+     * @param {object} compareOptions.blockOut blockout with x, y, width and height values
+     * @param {boolean} compareOptions.blockOutStatusBar blockout the statusbar yes or no, it will override the global
+     * @param {boolean} compareOptions.ignoreAntialiasing compare images an discard anti aliasing
+     * @param {boolean} compareOptions.ignoreColors Even though the images are in colour, the comparison wil compare 2 black/white images
      * @returns {Promise}
      * @private
      */
@@ -324,17 +320,17 @@ class protractorImageComparison {
      */
     _formatFileName(tag) {
         let defaults = {
-                'browserName': this.browserName,
-                'deviceName': this.deviceName,
-                'dpr': this.devicePixelRatio,
-                'height': this.height,
-                'logName': camelCase(this.logName),
-                'mobile': this._isMobile() && this.testInBrowser ? this.browserName : this._isMobile() ? 'app' : '',
-                'name': this.name,
-                'tag': tag,
-                'width': this.width
-            },
-            formatString = this.formatString;
+            'browserName': this.browserName,
+            'deviceName': this.deviceName,
+            'dpr': this.devicePixelRatio,
+            'height': this.height,
+            'logName': camelCase(this.logName),
+            'mobile': this._isMobile() && this.testInBrowser ? this.browserName : this._isMobile() ? 'app' : '',
+            'name': this.name,
+            'tag': tag,
+            'width': this.width
+        };
+        let formatString = this.formatString;
 
         defaults = this._mergeDefaultOptions(defaults, this.formatOptions);
 
@@ -569,7 +565,7 @@ class protractorImageComparison {
     _mergeDefaultOptions(optionsA, optionsB) {
         optionsB = (typeof optionsB === 'object') ? optionsB : {};
 
-        for (var option in optionsB) {
+        for (let option in optionsB) {
             if (optionsA.hasOwnProperty(option)) {
                 optionsA[option] = optionsB[option];
             }
@@ -616,7 +612,7 @@ class protractorImageComparison {
      * @returns {Promise}
      * @private
      */
-    _scrollToAndDetermineHeights(verticalCoordinate) {
+    _scrollToAndDetermineFullPageHeight(verticalCoordinate) {
         return browser.driver.executeAsyncScript(_scrollAndReturnFullPageHeight, verticalCoordinate, this.fullPageScrollTimeout);
 
         function _scrollAndReturnFullPageHeight(y, timeout, done) {
@@ -653,8 +649,9 @@ class protractorImageComparison {
         let actualFullPageHeight;
         let isLastScreenshot;
 
-        return this._scrollToAndDetermineHeights(verticalCoordinate)
+        return this._scrollToAndDetermineFullPageHeight(verticalCoordinate)
             .then(heights => {
+                // actualFullPageHeight = height;
                 console.log('\nScrolled to = ', verticalCoordinate);
                 actualFullPageHeight = heights.scrollHeight; // with dpr
                 this.innerHeight = heights.innerHeight;
@@ -676,13 +673,14 @@ class protractorImageComparison {
                     this.fullPageHeight = this.screenshotHeight;
                 } else {
                     // Value can be equal, or bigger due to for example lazyloading
-                    this.fullPageHeight = actualFullPageHeight - this.chromeShadowPadding;
+                    // this.fullPageHeight = actualFullPageHeight - this.chromeShadowPadding; // @todo: changed this for mobile
+                    this.fullPageHeight = actualFullPageHeight;
                 }
 
-                // Determine the crop height and postion from where to crop
-                isLastScreenshot = this.innerHeight * part > this.fullPageHeight;
-
                 if (this._isMobile() && this.nativeWebScreenshot) {
+                    // Determine the crop height and postion from where to crop
+                    isLastScreenshot = (this.innerHeight - this.chromeShadowPadding) * part > this.fullPageHeight;
+
                     // iOS + Android determine cropHeight for height
                     // iOS + Android determine cropTopPosition for y
                     // @todo, need to rethink the _getAndroid- and _getIOSPosition-, only give element coordinates back
@@ -701,9 +699,13 @@ class protractorImageComparison {
 
                     const statusAndAddressBarHeight = statusBarHeight + addressBarCurrentHeight;
 
-                    cropHeight = isLastScreenshot ? this.fullPageHeight - ((this.innerHeight - this.chromeShadowPadding) * (part - 1)) : this.innerHeight - this.chromeShadowPadding;
+                    // still got 6px extra black on the last screenshot, need to get this out, then it is good
+                    cropHeight = isLastScreenshot ? this.fullPageHeight - ((this.innerHeight - this.chromeShadowPadding) * (part - 1)) - this.chromeShadowPadding: this.innerHeight - this.chromeShadowPadding;
                     cropTopPosition = isLastScreenshot ? statusAndAddressBarHeight + this.innerHeight - cropHeight : statusAndAddressBarHeight + this.chromeShadowPadding;
                 } else {
+                    // Determine the crop height and postion from where to crop
+                    isLastScreenshot = this.innerHeight * part > this.fullPageHeight;
+
                     cropHeight = isLastScreenshot ? this.fullPageHeight - currentVerticalCoordinate : this.innerHeight;
                     cropTopPosition = isLastScreenshot ? this.innerHeight - cropHeight : 0;
                     // large image screenshot
@@ -728,7 +730,8 @@ class protractorImageComparison {
                     console.log('currentVerticalCoordinate = ', currentVerticalCoordinate);
                     console.log('verticalCoordinate = ', verticalCoordinate);
                     console.log('trying part =', part);
-                    console.log(`this.innerHeight * part > this.fullPageHeight = ${this.innerHeight * part} > ${this.fullPageHeight}`);
+                    console.log('isLastScreenshot = ', isLastScreenshot);
+                    console.log(`(this.innerHeight * part) - this.chromeShadowPadding > this.fullPageHeight = ${(this.innerHeight * part) - this.chromeShadowPadding} > ${this.fullPageHeight}`);
                     console.log('cropHeight = ', cropHeight);
                     console.log('cropTopPosition = ', cropTopPosition);
                     console.log('rectangles =', rectangles);
@@ -879,8 +882,8 @@ class protractorImageComparison {
      * @public
      */
     saveElement(element, tag, options) {
-        let saveOptions = options || [],
-            bufferedScreenshot;
+        let saveOptions = options || [];
+        let bufferedScreenshot;
 
         this.resizeDimensions = saveOptions.resizeDimensions ? saveOptions.resizeDimensions : this.resizeDimensions;
         this.disableCSSAnimation = saveOptions.disableCSSAnimation || saveOptions.disableCSSAnimation === false ? saveOptions.disableCSSAnimation : this.disableCSSAnimation;